Zen_Crypto_Trader/requirements.txt
ccxt
pandas
numpy
pyyaml
matplotlib
colorama
python-dotenv
